rename compiler to project
