+++ packages/compiler/src/project.ts
